API Documentation

Below is the API documentation for CatBoost.jl.

For a nice introduction to the package, see the examples.

CatBoost.Pool — Method

Pool(data; label=nothing, cat_features=nothing, text_features=nothing,
     pairs=nothing, delimiter='	', has_header=false, weight=nothing,
     group_id = nothing, group_weight=nothing, subgroup_id=nothing,
     pairs_weight=nothing, baseline=nothing, features_names=nothing,
     thread_count = -1) -> PyObject

Creates a Pool object holding training data and labels. data may also be passed as a keyword argument.

CatBoost.cv — Function

cv(pool::PyObject; kwargs...) -> DataFrame

Accepts a CatBoost.Pool positional argument to specify the training data, and keyword arguments to configure the settings. See the python documentation below for what keyword arguments are accepted.

Python documentation for catboost.cv

Cross-validate the CatBoost model.

Parameters
----------
pool : catboost.Pool
    Data to cross-validate on.

params : dict
    Parameters for CatBoost.
    CatBoost has many of parameters, all have default values.
    If  None, all params still defaults.
    If  dict, overriding some (or all) params.

dtrain : catboost.Pool or tuple (X, y)
    Synonym for pool parameter. Only one of these parameters should be set.

iterations : int
    Number of boosting iterations. Can be set in params dict.

num_boost_round : int
    Synonym for iterations. Only one of these parameters should be set.

fold_count : int, optional (default=3)
    The number of folds to split the dataset into.

nfold : int
    Synonym for fold_count.

type : string, optional (default='Classical')
    Type of cross-validation
    Possible values:
        - 'Classical'
        - 'Inverted'
        - 'TimeSeries'

inverted : bool, optional (default=False)
    Train on the test fold and evaluate the model on the training folds.

partition_random_seed : int, optional (default=0)
    Use this as the seed value for random permutation of the data.
    Permutation is performed before splitting the data for cross validation.
    Each seed generates unique data splits.

seed : int, optional
    Synonym for partition_random_seed. This parameter is deprecated. Use
    partition_random_seed instead.
    If both parameters are initialised partition_random_seed parameter is
    ignored.

shuffle : bool, optional (default=True)
    Shuffle the dataset objects before splitting into folds.

logging_level : string, optional (default=None)
    Possible values:
        - 'Silent'
        - 'Verbose'
        - 'Info'
        - 'Debug'

stratified : bool, optional (default=None)
    Perform stratified sampling. True for classification and False otherwise.

as_pandas : bool, optional (default=True)
    Return pd.DataFrame when pandas is installed.
    If False or pandas is not installed, return dict.

metric_period : int, [default=1]
    The frequency of iterations to print the information to stdout. The value should be a positive integer.

verbose : bool or int
    If verbose is bool, then if set to True, logging_level is set to Verbose,
    if set to False, logging_level is set to Silent.
    If verbose is int, it determines the frequency of writing metrics to output and
    logging_level is set to Verbose.

verbose_eval : bool or int
    Synonym for verbose. Only one of these parameters should be set.

plot : bool, optional (default=False)
    If True, draw train and eval error in Jupyter notebook

early_stopping_rounds : int
    Activates Iter overfitting detector with od_wait set to early_stopping_rounds.

save_snapshot : bool, [default=None]
    Enable progress snapshotting for restoring progress after crashes or interruptions

snapshot_file : string or pathlib.Path, [default=None]
    Learn progress snapshot file path, if None will use default filename

snapshot_interval: int, [default=600]
    Interval between saving snapshots (seconds)

metric_update_interval: float, [default=0.5]
    Interval between updating metrics (seconds)

folds: generator or iterator of (train_idx, test_idx) tuples, scikit-learn splitter object or None, optional (default=None)
    If generator or iterator, it should yield the train and test indices for each fold.
    If object, it should be one of the scikit-learn splitter classes
    (https://scikit-learn.org/stable/modules/classes.html#splitter-classes)
    and have ``split`` method.
    if folds is not None, then all of fold_count, shuffle, partition_random_seed, inverted are None

return_models: bool, optional (default=False)
    if True, return a list of models fitted for each CV fold

log_cout: output stream or callback for logging

log_cerr: error stream or callback for logging

Returns
-------
cv results : pandas.core.frame.DataFrame with cross-validation results
    columns are: test-error-mean  test-error-std  train-error-mean  train-error-std
cv models : list of trained models, if return_models=True

CatBoost.to_catboost — Method

to_catboost(arg)

to_catboost is called on each argument passed to fit, predict, predict_proba, and cv to allow customization of the conversion of Julia types to python types. If to_catboost emits a Julia type, then PyCall will try to convert it appropriately (automatically).

By default, to_catboost simply checks if the argument satisfies Tables.istable(arg), and if so, it outputs a corresponding pandas table, and otherwise passes it on.

To customize the conversion for custom types, provide a method for this function.